modified

+++ src/main/java/iiitd/harsh22199/Event.java
@@ -5,7 +5,7 @@ public interface Event {
     public void manageEvent();
     public void setEvent_name(String event_name);
     public String getEvent_name();
-    public float getEvent_price();
-    public void setEvent_price(float event_price);
+    public double getEvent_price();
+    public void setEvent_price(double event_price);
     public String toString();
 }
